+++ app/layout.js
@@ -40,7 +40,7 @@ export default function RootLayout({ children }) {
         </header>
         {children}
         <div className="bg-[#474a47]  text-white/60 border-dashed h-fit border-t-12 flex flex-col justify-center items-center border-[#232323]/80 b p-3 shadow-2xl shadow-amber-50">
-       <div className="hidden ">   <Menu/></div>
+       <div >   <Link href="PaS" className="text-green-100 flex gap-2 font-sans opacity-60 px-3 m-3 py-1  border-green border ring-4 rounded-xl  ring-gray-800  font-bold italic text-balance bg-gray-500 mt-5 hover:scale-110 cursor-pointer min-md:hidden">Home</Link></div>
           <div className="rounded-2xl border p-1 border-[#454545] bg-[#373937]/80">© 2025{new Date().getFullYear() == 2025? "":new Date().getFullYear() } Sleepy Panda All rights Reserved</div>
         </div>
       </body>
